updated combined trade tests
